org.ybygjy.basic/network/src/main/java/org/ybygjy/basic/network/rpc/serialize/ObjectSerializationService.java
package org.ybygjy.basic.network.rpc.serialize;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.IOException;
import java.io.ObjectInputStream;
import java.io.ObjectOutputStream;
import java.io.Serializable;

/**
 * 对象序列化服务
 * @author WangYanCheng
 * @version 2016年9月11日
 */
public class ObjectSerializationService {
    public byte[] serializeObject(Serializable obj) throws IOException {
        ByteArrayOutputStream baos = new ByteArrayOutputStream();
        ObjectOutputStream oos = new ObjectOutputStream(baos);
        oos.writeObject(obj);
        return baos.toByteArray();
    }
    public Object readObject(byte[] objDataArr) throws IOException, ClassNotFoundException {
        ObjectInputStream ois = new ObjectInputStream(new ByteArrayInputStream(objDataArr));
        Object object = ois.readObject();
        return object;
    }
}
